note: big todo comment add re: data migrations
I added 3 new puzzle sets, but they aren't showing up in game b/c the
save-game code only fetches puzzle_sets that were saved.

I'm realizing it'd be better to save user events than to save some
specific state of the data - it'll be simpler to have a list of events
like 'completed-puzzle-set {set=FALL}' that can be used to recreate a
working state - that'll be able to be reapplied to updated data, which
should reduce migration complexity. Plus those events will probably be
useful for other things, like achievements and records and such.