feat: working event-y save-game feature
The final change was to drop the reference property in favor of a string
entity_id. It's much weaker but it actually works. I suspect the
deserialized reference type was coming out as a string? Not really sure
what happened. A lingering concern is events failing to deserialize -
deleting the property seemed to break the deserialization :/

Anyway, happy to move forward with this for now.