fix: remove "Calls to action" itch links in credits
Steam rejected the build because i included text (unclickable!) links to
some of the free assets used to make this game. Calling these 'calls to
action' is quiet a stretch, and the language in the rejection says it's
ok if it goes through steam wallet... for free assets? Eh, bunch of
corpo bullshit.