feat: extend themes to support a list of optional music tracks
Only the first will be played, but a list seems to be better long term.
This allows tracks to be referenced that may not exist in the repo (e.g.
during CI builds). If no track in the list is present, the
background_music property is used as a fallback.