fix: more focus tweaks
The worldmap is still loading without grabbing focus - I'm not sure
what's the deal, so here's some more aggressive focus grabbing. This is
super annoying b/c it completely breaks the game if the user doesn't
have a mouse/screen-input to fix it. Probably should impl some fallback
so that arbitrary inputs check and find a focus if necessary.