feat: unlock puzzles based on puzzles completed
Rather than a linear one-world-at-a-time approach, this allows for
unlocking the next world after fewer in the previous one.

Rearranges the on-win logic in the game_scene, pulling some stat calcs
out of the store functions. Maybe the achievements belong in the store?
but it doesn't seem like the ui-flow/jumbotrons do.