feat: rework puzzle-win logic
performs calcs to properly mark puzzle-sets complete (important now that
puzzles can be skipped). Updates completion of the final puzzle with a
more specific note, and drops the intermediate puzzle-complete
events (puzzles are broken up by puzzle-unlock events now).

Perhaps we want to jump to the next incomplete level, or provide a menu
for selecting the next puzzle if the next one is already complete?