chore: misc log clean up and what not
Things back to working, now with 'GameDef'!

Hurting for a bit more of a refactor - we're relying on PuzzleSet
entities for game state/save data, when we probably want something less
global, and more game mode aware.

Plus i really want a PuzzleStore that has reasonable apis to pull
from. (But in what game-mode context? Is PuzzleStore game-mode dependent
or agnostic?)