fix: serialize parsedGame on PuzzleSetData
Exported games were crashing b/c the .puzz files were not parsable in
exported builds - this serialization is a step i hesitated to take when
first implementing the .puzz importer, but it turned out to be pretty
easy. We could probably serialize the PuzzleDefs and maybe the analysis
too... hmmmmmm.