feat: introduce Events.puzzle_node, GameSounds as first consumer
Moving away from inheritance here - the puzzleNodeExtenders are a little
bit drier, but they are not composable with other things (e.g. UI
buttons for changing the theme).

Maybe this setup could be dried up with a helper expecting a
ready_event, a setup_node, and a list_nodes input?