feat: intro 'Random' mode, split out and reuse classic hud controls
Random Mode is like Treadmill, but it pulls a random puzzle_def instead
of repeating.

The HUD had all the UI in it - this breaks out the control buttons/hints
into a reusable comp. The HUD is tied to classic mode's semantics, and
we'll want some other toggles in the Gym to toggle things on/off.